test: 合并单元测试工程
