+++ BootstrapBlazor.sln
@@ -144,17 +144,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BootstrapBlazor.DockView", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BootstrapBlazor.IconPark", "src\Extensions\Components\BootstrapBlazor.IconPark\BootstrapBlazor.IconPark.csproj", "{6BDE87D7-62D9-4650-95F5-B60225BAFA68}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UniTestIconPark", "test\UniTestIconPark\UniTestIconPark.csproj", "{D46E8593-641F-4C90-9D95-3ABF41A327FB}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BootstrapBlazor.Html2Pdf.iText", "src\Extensions\Components\BootstrapBlazor.Html2Pdf.iText\BootstrapBlazor.Html2Pdf.iText.csproj", "{AF077995-AFE6-4065-9B3C-CC0B79D9D179}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BootstrapBlazor.DriverJs", "src\Extensions\Components\BootstrapBlazor.DriverJs\BootstrapBlazor.DriverJs.csproj", "{B1F29F72-E0C0-41F6-A38E-A02B775C2FE7}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BootstrapBlazor.AntDesignIcon", "src\Extensions\Components\BootstrapBlazor.AntDesignIcon\BootstrapBlazor.AntDesignIcon.csproj", "{7C09E567-AEEA-4ED1-BFF5-D06479937DAC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UniTestAntDesignIcon", "test\UniTestAntDesignIcon\UniTestAntDesignIcon.csproj", "{DC9B84B2-AF95-4424-AFCE-D86040AFC5AF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BootstrapBlazor.ElementIcon", "src\Extensions\Components\BootstrapBlazor.ElementIcon\BootstrapBlazor.ElementIcon.csproj", "{481653B0-B096-4F72-BA64-83F52E97E45B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BootstrapBlazor.ElementIcon", "src\Extensions\Components\BootstrapBlazor.ElementIcon\BootstrapBlazor.ElementIcon.csproj", "{481653B0-B096-4F72-BA64-83F52E97E45B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UniTestSvgIcon", "test\UniTestSvgIcon\UniTestSvgIcon.csproj", "{01DA4687-0973-4CA6-97F5-04E0B8DD1B13}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -340,10 +338,6 @@ Global
 		{6BDE87D7-62D9-4650-95F5-B60225BAFA68}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6BDE87D7-62D9-4650-95F5-B60225BAFA68}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6BDE87D7-62D9-4650-95F5-B60225BAFA68}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D46E8593-641F-4C90-9D95-3ABF41A327FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D46E8593-641F-4C90-9D95-3ABF41A327FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D46E8593-641F-4C90-9D95-3ABF41A327FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D46E8593-641F-4C90-9D95-3ABF41A327FB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AF077995-AFE6-4065-9B3C-CC0B79D9D179}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AF077995-AFE6-4065-9B3C-CC0B79D9D179}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AF077995-AFE6-4065-9B3C-CC0B79D9D179}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -356,14 +350,14 @@ Global
 		{7C09E567-AEEA-4ED1-BFF5-D06479937DAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7C09E567-AEEA-4ED1-BFF5-D06479937DAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7C09E567-AEEA-4ED1-BFF5-D06479937DAC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DC9B84B2-AF95-4424-AFCE-D86040AFC5AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DC9B84B2-AF95-4424-AFCE-D86040AFC5AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DC9B84B2-AF95-4424-AFCE-D86040AFC5AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DC9B84B2-AF95-4424-AFCE-D86040AFC5AF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{481653B0-B096-4F72-BA64-83F52E97E45B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{481653B0-B096-4F72-BA64-83F52E97E45B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{481653B0-B096-4F72-BA64-83F52E97E45B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{481653B0-B096-4F72-BA64-83F52E97E45B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{01DA4687-0973-4CA6-97F5-04E0B8DD1B13}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{01DA4687-0973-4CA6-97F5-04E0B8DD1B13}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{01DA4687-0973-4CA6-97F5-04E0B8DD1B13}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{01DA4687-0973-4CA6-97F5-04E0B8DD1B13}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -421,12 +415,11 @@ Global
 		{CA517F07-F3BA-424B-BACD-3D83122B38CD} = {CD062AB6-244D-402A-8F33-C37DAC5856CC}
 		{EB33F94D-BCA7-4B85-98E5-C093D798C804} = {CD062AB6-244D-402A-8F33-C37DAC5856CC}
 		{6BDE87D7-62D9-4650-95F5-B60225BAFA68} = {CD062AB6-244D-402A-8F33-C37DAC5856CC}
-		{D46E8593-641F-4C90-9D95-3ABF41A327FB} = {7C1D79F1-87BC-42C1-BD5A-CDE4044AC1BD}
 		{AF077995-AFE6-4065-9B3C-CC0B79D9D179} = {CD062AB6-244D-402A-8F33-C37DAC5856CC}
 		{B1F29F72-E0C0-41F6-A38E-A02B775C2FE7} = {CD062AB6-244D-402A-8F33-C37DAC5856CC}
 		{7C09E567-AEEA-4ED1-BFF5-D06479937DAC} = {CD062AB6-244D-402A-8F33-C37DAC5856CC}
-		{DC9B84B2-AF95-4424-AFCE-D86040AFC5AF} = {7C1D79F1-87BC-42C1-BD5A-CDE4044AC1BD}
 		{481653B0-B096-4F72-BA64-83F52E97E45B} = {CD062AB6-244D-402A-8F33-C37DAC5856CC}
+		{01DA4687-0973-4CA6-97F5-04E0B8DD1B13} = {7C1D79F1-87BC-42C1-BD5A-CDE4044AC1BD}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {0DCB0756-34FA-4FD0-AE1D-D3F08B5B3A6B}
