+++ test/UniTestSvgIcon/UnitTest.cs
@@ -11,7 +11,7 @@ namespace UniTestIconPark;
 public partial class UnitTest
 {
     [Fact]
-    public void Build()
+    public void IconPark_Ok()
     {
         var services = new ServiceCollection();
         services.AddBootstrapBlazor();
@@ -68,8 +68,76 @@ public partial class UnitTest
         writer.Close();
     }
 
+    [Theory]
+    [InlineData("filled")]
+    [InlineData("outlined")]
+    [InlineData("twotone")]
+    public void AntDesignIcon_Ok(string category)
+    {
+        var services = new ServiceCollection();
+        services.AddBootstrapBlazor();
+        var provider = services.BuildServiceProvider();
+        var zipService = provider.GetRequiredService<IZipArchiveService>();
+
+        var root = AppContext.BaseDirectory;
+        var downloadFile = Path.Combine(root, $"{category}.zip");
+        Assert.True(File.Exists(downloadFile));
+
+        var downloadFolder = Path.Combine(root, category);
+        if (Directory.Exists(downloadFile))
+        {
+            Directory.Delete(downloadFile, true);
+        }
+        zipService.ExtractToDirectory(downloadFile, downloadFolder, true);
+
+        var folder = new DirectoryInfo(downloadFolder);
+
+        // 处理 List 文件
+        var iconListFile = Path.Combine(root, $"../../../AntDesignIconList_{category}.razor");
+        if (File.Exists(iconListFile))
+        {
+            File.Delete(iconListFile);
+        }
+
+        // 处理 svg 文件
+        var svgFile = Path.Combine(root, $"../../../{category}.svg");
+        if (File.Exists(svgFile))
+        {
+            File.Delete(svgFile);
+        }
+        using var listWriter = new StreamWriter(File.OpenWrite(iconListFile));
+        using var writer = new StreamWriter(File.OpenWrite(svgFile));
+        writer.WriteLine("<svg xmlns=\"http://www.w3.org/2000/svg\">");
+        foreach (var icon in folder.EnumerateFiles())
+        {
+            var id = Path.GetFileNameWithoutExtension(icon.Name);
+            using var reader = new StreamReader(icon.OpenRead());
+            var data = reader.ReadToEnd();
+            reader.Close();
+
+            // find <svg
+            var index = data.IndexOf("<svg ");
+            if (index > -1)
+            {
+                data = data[index..];
+            }
+            index = data.IndexOf(">");
+            if (index > -1)
+            {
+                data = data[(index + 1)..];
+            }
+            var target = data.Replace("</svg>", "").Trim();
+            target = $"    <symbol viewBox=\"0 0 1024 1024\" id=\"{id}\">{target}</symbol>";
+            writer.WriteLine(target);
+
+            listWriter.WriteLine($"<AntDesignIcon Name=\"{id}\"></AntDesignIcon>");
+        }
+        writer.WriteLine("</svg>");
+        writer.Close();
+    }
+
     [Fact]
-    public void Element_Ok()
+    public void ElementIcon_Ok()
     {
         var category = "element";
         var services = new ServiceCollection();
