+++ test/UniTestSvgIcon/element.zip
