+++ test/UniTestSvgIcon/filled.zip
