+++ test/UniTestSvgIcon/outlined.zip
